Rename ConnectLoginButtonWithProgressBar -> ConnectLoginButton
