+++ connect-id-example/src/main/res/layout/activity_sign_in.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     tools:context="com.telenor.connect.connectidexample.SignInActivity">
 
-    <com.telenor.connect.ui.ConnectLoginButtonWithProgressBar
+    <com.telenor.connect.ui.ConnectLoginButton
         android:id="@+id/login_button"
         android:layout_centerInParent="true"
         android:layout_width="wrap_content"
